+++ iDo/Classes/Util/GATTUtils.swift
@@ -15,22 +15,22 @@ let BLE_UUID_TEMPERATURE_MEASUREMENT = "2A1C" // Temperature Measurement
 let kCharacteristicUUID = "2A1E"
 
 /** 处理蓝牙传来的data */
-func calculateTemperature(data: NSData) -> Float {
+func calculateTemperature(data: NSData) -> Double {
     var bytes = [UInt8](count: 5, repeatedValue: 0)
     data.getBytes(&bytes, length: 5)
-    let exponent = Float(-(255 - bytes[4]) - 1) // 我自己都忘了为什么这么写，总之很对就是了
+    let exponent = Double(-(255 - bytes[4]) - 1) // 我自己都忘了为什么这么写，总之很对就是了
     let mantissa = Int32(bytes[3]) << 16 | Int32(bytes[2]) << 8 | Int32(bytes[1])
-    return Float(mantissa) * pow(10.0, exponent)
+    return Double(mantissa) * pow(10.0, exponent)
 }
 
-func calculateAdvTemperature(data: NSData) -> Float {
+func calculateAdvTemperature(data: NSData) -> Double {
     var bytes = [UInt8](count: 4, repeatedValue: 0)
     data.getBytes(&bytes, length: 4)
-    let exponent = Float(-(255 - bytes[3]) - 1) // 我自己都忘了为什么这么写，总之很对就是了
+    let exponent = Double(-(255 - bytes[3]) - 1) // 我自己都忘了为什么这么写，总之很对就是了
     let mantissa = Int32(bytes[2]) << 16 | Int32(bytes[1]) << 8 | Int32(bytes[0])
-    return Float(mantissa) * pow(10.0, exponent)
+    return Double(mantissa) * pow(10.0, exponent)
 }
 
-func transformTemperature(value: Float) -> Float {
-    return Float(32) + value * 1.8
+func transformTemperature(value: Double) -> Double {
+    return Double(32) + value * 1.8
 }
